requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0
opencv-python>=4.8.0
psutil>=5.9.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.1.0
python-dotenv>=1.0.0
Pillow>=9.5.0
pywhatkit>=5.4.0
twilio>=8.0.0
google-search-results>=2.4.0
mediapipe>=0.10.0
